+++ api.h
@@ -54,6 +54,9 @@
 #define REAPERAPI_WANT_Undo_OnStateChangeEx
 #define REAPERAPI_WANT_GetMainHwnd
 
+// Project state change polling (Layers track-order sync, shadow-map staleness)
+#define REAPERAPI_WANT_GetProjectStateChangeCount
+
 // Plugin registration (timer, projectconfig, command_id, gaccel, hookcommand)
 #define REAPERAPI_WANT_plugin_register